pyproject.toml
[project]
name = "vision-robotics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.21",
    "pillow>=11.3.0",
    "pyserial>=3.5",
]
